@@ -1,19 +1,28 @@
 module github.com/npolizotis/saml
 
-go 1.16
+go 1.19
 
 require (
-	github.com/beevik/etree v1.1.0
+	github.com/beevik/etree v1.2.0
 	github.com/crewjam/httperr v0.2.0
 	github.com/dchest/uniuri v1.2.0
-	github.com/golang-jwt/jwt/v4 v4.4.3
-	github.com/google/go-cmp v0.5.9
+	github.com/golang-jwt/jwt/v4 v4.5.0
+	github.com/google/go-cmp v0.6.0
 	github.com/kr/pretty v0.3.1
 	github.com/mattermost/xml-roundtrip-validator v0.1.0
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/russellhaering/goxmldsig v1.3.0
-	github.com/stretchr/testify v1.8.1
+	github.com/russellhaering/goxmldsig v1.4.0
+	github.com/stretchr/testify v1.8.4
 	github.com/zenazn/goji v1.0.1
-	golang.org/x/crypto v0.0.0-20220128200615-198e4374d7ed
+	golang.org/x/crypto v0.14.0
 	gotest.tools v2.2.0+incompatible
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/jonboulle/clockwork v0.2.2 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rogpeppe/go-internal v1.9.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
